+++ tests/text/helpers.py
@@ -11,18 +11,18 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import os
 import pickle
 import sys
+from enum import Enum, unique
 from functools import partial
-from typing import Any, Callable, Dict, Optional, Sequence
+from typing import Any, Callable, Sequence, Union
 
-import numpy as np
 import pytest
 import torch
-from torch import Tensor, tensor
-from torch.multiprocessing import Pool, set_start_method
+from torch import Tensor
+from torch.multiprocessing import set_start_method
 
+from tests.helpers.testers import MetricTester, _assert_allclose, _assert_requires_grad, _assert_tensor
 from torchmetrics import Metric
 
 try:
@@ -30,82 +30,22 @@ try:
 except RuntimeError:
     pass
 
-NUM_PROCESSES = 2
-NUM_BATCHES = 10
-BATCH_SIZE = 32
-NUM_CLASSES = 5
-EXTRA_DIM = 3
-THRESHOLD = 0.5
 
-MAX_PORT = 8100
-START_PORT = 8088
-CURRENT_PORT = START_PORT
+@unique
+class INPUT_ORDER(Enum):
+    PREDS_FIRST = 1
+    TARGETS_FIRST = 2
 
 
-def setup_ddp(rank, world_size):
-    """Setup ddp environment."""
-    global CURRENT_PORT
-
-    os.environ["MASTER_ADDR"] = "localhost"
-    os.environ["MASTER_PORT"] = str(CURRENT_PORT)
-
-    CURRENT_PORT += 1
-    if CURRENT_PORT > MAX_PORT:
-        CURRENT_PORT = START_PORT
-
-    if torch.distributed.is_available() and sys.platform not in ("win32", "cygwin"):
-        torch.distributed.init_process_group("gloo", rank=rank, world_size=world_size)
-
-
-def _assert_allclose(pl_result: Any, sk_result: Any, atol: float = 1e-8, key: Optional[str] = None) -> None:
-    """Utility function for recursively asserting that two results are within a certain tolerance."""
-    # single output compare
-    if isinstance(pl_result, Tensor):
-        assert np.allclose(pl_result.cpu().numpy(), sk_result, atol=atol, equal_nan=True)
-    # multi output compare
-    elif isinstance(pl_result, Sequence):
-        for pl_res, sk_res in zip(pl_result, sk_result):
-            _assert_allclose(pl_res, sk_res, atol=atol)
-    elif isinstance(pl_result, Dict):
-        if key is None:
-            raise KeyError("Provide Key for Dict based metric results.")
-        assert np.allclose(pl_result[key].detach().cpu().numpy(), sk_result, atol=atol, equal_nan=True)
-    else:
-        raise ValueError("Unknown format for comparison")
-
-
-def _assert_tensor(pl_result: Any, key: Optional[str] = None) -> None:
-    """Utility function for recursively checking that some input only consists of torch tensors."""
-    if isinstance(pl_result, Sequence):
-        for plr in pl_result:
-            _assert_tensor(plr)
-    elif isinstance(pl_result, Dict):
-        if key is None:
-            raise KeyError("Provide Key for Dict based metric results.")
-        assert isinstance(pl_result[key], Tensor)
-    else:
-        assert isinstance(pl_result, Tensor)
-
-
-def _assert_requires_grad(metric: Metric, pl_result: Any, key: Optional[str] = None) -> None:
-    """Utility function for recursively asserting that metric output is consistent with the `is_differentiable`
-    attribute."""
-    if isinstance(pl_result, Sequence):
-        for plr in pl_result:
-            _assert_requires_grad(metric, plr, key=key)
-    elif isinstance(pl_result, Dict):
-        if key is None:
-            raise KeyError("Provide Key for Dict based metric results.")
-        assert metric.is_differentiable == pl_result[key].requires_grad
-    else:
-        assert metric.is_differentiable == pl_result.requires_grad
+TEXT_METRIC_INPUT = Union[Sequence[str], Sequence[Sequence[str]], Sequence[Sequence[Sequence[str]]]]
+NUM_BATCHES = 2
 
 
 def _class_test(
     rank: int,
     worldsize: int,
-    preds: Tensor,
-    target: Tensor,
+    preds: TEXT_METRIC_INPUT,
+    targets: TEXT_METRIC_INPUT,
     metric_class: Metric,
     sk_metric: Callable,
     dist_sync_on_step: bool,
@@ -116,6 +56,8 @@ def _class_test(
     device: str = "cpu",
     fragment_kwargs: bool = False,
     check_scriptable: bool = True,
+    input_order: INPUT_ORDER = INPUT_ORDER.PREDS_FIRST,
+    key: str = None,
     **kwargs_update: Any,
 ):
     """Utility function doing the actual comparison between lightning class metric and reference metric.
@@ -123,8 +65,8 @@ def _class_test(
     Args:
         rank: rank of current process
         worldsize: number of processes
-        preds: torch tensor with predictions
-        target: torch tensor with targets
+        preds: Sequence of predicted tokens or predicted sentences
+        targets: Sequence of target tokens or target sentences
         metric_class: lightning metric class that should be tested
         sk_metric: callable function that is used for comparison
         dist_sync_on_step: bool, if true will synchronize metric state across
@@ -135,9 +77,12 @@ def _class_test(
         check_batch: bool, if true will check if the metric is also correctly
             calculated across devices for each batch (and not just at the end)
         device: determine which device to run on, either 'cuda' or 'cpu'
-        fragment_kwargs: whether tensors in kwargs should be divided as `preds` and `target` among processes
+        fragment_kwargs: whether tensors in kwargs should be divided as `preds` and `targets` among processes
+        input_order: Define the ordering for the preds and targets positional arguments.
+        key: The key passed onto the `_assert_allclose` to compare the respective metric from the Dict output against
+            the sk_metric.
         kwargs_update: Additional keyword arguments that will be passed with preds and
-            target when running update on the metric.
+            targets when running update on the metric.
     """
     if not metric_args:
         metric_args = {}
@@ -153,8 +98,6 @@ def _class_test(
 
     # move to device
     metric = metric.to(device)
-    preds = preds.to(device)
-    target = target.to(device)
     kwargs_update = {k: v.to(device) if isinstance(v, Tensor) else v for k, v in kwargs_update.items()}
 
     # verify metrics work after being loaded from pickled state
@@ -164,95 +107,126 @@ def _class_test(
     for i in range(rank, NUM_BATCHES, worldsize):
         batch_kwargs_update = {k: v[i] if isinstance(v, Tensor) else v for k, v in kwargs_update.items()}
 
-        batch_result = metric(preds[i], target[i], **batch_kwargs_update)
+        if input_order == INPUT_ORDER.PREDS_FIRST:
+            batch_result = metric(preds[i], targets[i], **batch_kwargs_update)
+        elif input_order == INPUT_ORDER.TARGETS_FIRST:
+            batch_result = metric(targets[i], preds[i], **batch_kwargs_update)
 
         if metric.dist_sync_on_step and check_dist_sync_on_step and rank == 0:
-            ddp_preds = torch.cat([preds[i + r] for r in range(worldsize)]).cpu()
-            ddp_target = torch.cat([target[i + r] for r in range(worldsize)]).cpu()
+            # Concatenation of Sequence of strings
+            ddp_preds = type(preds)()
+            ddp_targets = type(targets)()
+            for r in range(worldsize):
+                ddp_preds = ddp_preds + preds[i + r]
+                ddp_targets = ddp_targets + targets[i + r]
             ddp_kwargs_upd = {
                 k: torch.cat([v[i + r] for r in range(worldsize)]).cpu() if isinstance(v, Tensor) else v
                 for k, v in (kwargs_update if fragment_kwargs else batch_kwargs_update).items()
             }
 
-            sk_batch_result = sk_metric(ddp_preds, ddp_target, **ddp_kwargs_upd)
-            _assert_allclose(batch_result, sk_batch_result, atol=atol)
+            if input_order == INPUT_ORDER.PREDS_FIRST:
+                sk_batch_result = sk_metric(ddp_preds, ddp_targets, **ddp_kwargs_upd)
+            elif input_order == INPUT_ORDER.TARGETS_FIRST:
+                sk_batch_result = sk_metric(ddp_targets, ddp_preds, **ddp_kwargs_upd)
+            _assert_allclose(batch_result, sk_batch_result, atol=atol, key=key)
 
         elif check_batch and not metric.dist_sync_on_step:
             batch_kwargs_update = {
                 k: v.cpu() if isinstance(v, Tensor) else v
                 for k, v in (batch_kwargs_update if fragment_kwargs else kwargs_update).items()
             }
-            sk_batch_result = sk_metric(preds[i].cpu(), target[i].cpu(), **batch_kwargs_update)
-            _assert_allclose(batch_result, sk_batch_result, atol=atol)
+            if input_order == INPUT_ORDER.PREDS_FIRST:
+                sk_batch_result = sk_metric(preds[i], targets[i], **batch_kwargs_update)
+            elif input_order == INPUT_ORDER.TARGETS_FIRST:
+                sk_batch_result = sk_metric(targets[i], preds[i], **batch_kwargs_update)
+
+            _assert_allclose(batch_result, sk_batch_result, atol=atol, key=key)
 
     # check on all batches on all ranks
     result = metric.compute()
-    _assert_tensor(result)
+    _assert_tensor(result, key=key)
 
-    total_preds = torch.cat([preds[i] for i in range(NUM_BATCHES)]).cpu()
-    total_target = torch.cat([target[i] for i in range(NUM_BATCHES)]).cpu()
+    # Concatenation of Sequence of strings
+    total_preds = type(preds)()
+    total_targets = type(targets)()
+    for i in range(NUM_BATCHES):
+        total_preds = total_preds + preds[i]
+        total_targets = total_targets + targets[i]
     total_kwargs_update = {
         k: torch.cat([v[i] for i in range(NUM_BATCHES)]).cpu() if isinstance(v, Tensor) else v
         for k, v in kwargs_update.items()
     }
-    sk_result = sk_metric(total_preds, total_target, **total_kwargs_update)
+    if input_order == INPUT_ORDER.PREDS_FIRST:
+        sk_result = sk_metric(total_preds, total_targets, **total_kwargs_update)
+    elif input_order == INPUT_ORDER.TARGETS_FIRST:
+        sk_result = sk_metric(total_targets, total_preds, **total_kwargs_update)
 
     # assert after aggregation
-    _assert_allclose(result, sk_result, atol=atol)
+    _assert_allclose(result, sk_result, atol=atol, key=key)
 
 
 def _functional_test(
-    preds: Tensor,
-    target: Tensor,
+    preds: TEXT_METRIC_INPUT,
+    targets: TEXT_METRIC_INPUT,
     metric_functional: Callable,
     sk_metric: Callable,
     metric_args: dict = None,
     atol: float = 1e-8,
     device: str = "cpu",
     fragment_kwargs: bool = False,
+    input_order: INPUT_ORDER = INPUT_ORDER.PREDS_FIRST,
+    key: str = None,
     **kwargs_update,
 ):
     """Utility function doing the actual comparison between lightning functional metric and reference metric.
 
     Args:
         preds: torch tensor with predictions
-        target: torch tensor with targets
+        targets: torch tensor with targets
         metric_functional: lightning metric functional that should be tested
         sk_metric: callable function that is used for comparison
         metric_args: dict with additional arguments used for class initialization
         device: determine which device to run on, either 'cuda' or 'cpu'
-        fragment_kwargs: whether tensors in kwargs should be divided as `preds` and `target` among processes
+        fragment_kwargs: whether tensors in kwargs should be divided as `preds` and `targets` among processes
+        input_order: Define the ordering for the preds and targets positional arguments.
+        key: The key passed onto the `_assert_allclose` to compare the respective metric from the Dict output against
+            the sk_metric.
         kwargs_update: Additional keyword arguments that will be passed with preds and
-            target when running update on the metric.
+            targets when running update on the metric.
     """
     if not metric_args:
         metric_args = {}
 
     metric = partial(metric_functional, **metric_args)
 
-    # move to device
-    preds = preds.to(device)
-    target = target.to(device)
+    # Move to device
     kwargs_update = {k: v.to(device) if isinstance(v, Tensor) else v for k, v in kwargs_update.items()}
 
     for i in range(NUM_BATCHES):
         extra_kwargs = {k: v[i] if isinstance(v, Tensor) else v for k, v in kwargs_update.items()}
-        lightning_result = metric(preds[i], target[i], **extra_kwargs)
+        if input_order == INPUT_ORDER.PREDS_FIRST:
+            lightning_result = metric(preds[i], targets[i], **extra_kwargs)
+        elif input_order == INPUT_ORDER.TARGETS_FIRST:
+            lightning_result = metric(targets[i], preds[i], **extra_kwargs)
+
         extra_kwargs = {
             k: v.cpu() if isinstance(v, Tensor) else v
             for k, v in (extra_kwargs if fragment_kwargs else kwargs_update).items()
         }
-        sk_result = sk_metric(preds[i].cpu(), target[i].cpu(), **extra_kwargs)
+        if input_order == INPUT_ORDER.PREDS_FIRST:
+            sk_result = sk_metric(preds[i], targets[i], **extra_kwargs)
+        elif input_order == INPUT_ORDER.TARGETS_FIRST:
+            sk_result = sk_metric(targets[i], preds[i], **extra_kwargs)
 
         # assert its the same
-        _assert_allclose(lightning_result, sk_result, atol=atol)
+        _assert_allclose(lightning_result, sk_result, atol=atol, key=key)
 
 
 def _assert_half_support(
     metric_module: Metric,
     metric_functional: Callable,
-    preds: Tensor,
-    target: Tensor,
+    preds: TEXT_METRIC_INPUT,
+    targets: TEXT_METRIC_INPUT,
     device: str = "cpu",
     **kwargs_update,
 ):
@@ -262,13 +236,13 @@ def _assert_half_support(
         metric_module: the metric module to test
         metric_functional: the metric functional to test
         preds: torch tensor with predictions
-        target: torch tensor with targets
+        targets: torch tensor with targets
         device: determine device, either "cpu" or "cuda"
         kwargs_update: Additional keyword arguments that will be passed with preds and
-                target when running update on the metric.
+                targets when running update on the metric.
     """
-    y_hat = preds[0].half().to(device) if preds[0].is_floating_point() else preds[0].to(device)
-    y = target[0].half().to(device) if target[0].is_floating_point() else target[0].to(device)
+    y_hat = preds[0]
+    y = targets[0]
     kwargs_update = {
         k: (v[0].half() if v.is_floating_point() else v[0]).to(device) if isinstance(v, Tensor) else v
         for k, v in kwargs_update.items()
@@ -278,72 +252,62 @@ def _assert_half_support(
     _assert_tensor(metric_functional(y_hat, y, **kwargs_update))
 
 
-class MetricTester:
+class TextTester(MetricTester):
     """Class used for efficiently run alot of parametrized tests in ddp mode. Makes sure that ddp is only setup
     once and that pool of processes are used for all tests.
 
-    All tests should subclass from this and implement a new method called     `test_metric_name` where the method
-    `self.run_metric_test` is called inside.
+    All tests for text metrics should subclass from this and implement a new method called `test_metric_name` where the
+    method `self.run_metric_test` is called inside.
     """
-
-    atol = 1e-8
-
-    def setup_class(self):
-        """Setup the metric class.
-
-        This will spawn the pool of workers that are used for metric testing and setup_ddp
-        """
-
-        self.poolSize = NUM_PROCESSES
-        self.pool = Pool(processes=self.poolSize)
-        self.pool.starmap(setup_ddp, [(rank, self.poolSize) for rank in range(self.poolSize)])
-
-    def teardown_class(self):
-        """Close pool of workers."""
-        self.pool.close()
-        self.pool.join()
 
     def run_functional_metric_test(
         self,
-        preds: Tensor,
-        target: Tensor,
+        preds: TEXT_METRIC_INPUT,
+        targets: TEXT_METRIC_INPUT,
         metric_functional: Callable,
         sk_metric: Callable,
         metric_args: dict = None,
         fragment_kwargs: bool = False,
+        input_order: INPUT_ORDER = INPUT_ORDER.PREDS_FIRST,
+        key: str = None,
         **kwargs_update,
     ):
         """Main method that should be used for testing functions. Call this inside testing method.
 
         Args:
             preds: torch tensor with predictions
-            target: torch tensor with targets
+            targets: torch tensor with targets
             metric_functional: lightning metric class that should be tested
             sk_metric: callable function that is used for comparison
             metric_args: dict with additional arguments used for class initialization
-            fragment_kwargs: whether tensors in kwargs should be divided as `preds` and `target` among processes
+            fragment_kwargs: whether tensors in kwargs should be divided as `preds` and `targets` among processes
+            input_order: Define the ordering for the preds and targets positional arguments.
+            key: The key passed onto the `_assert_allclose` to compare the respective metric from the Dict output
+                against the sk_metric.
             kwargs_update: Additional keyword arguments that will be passed with preds and
-                target when running update on the metric.
+                targets when running update on the metric.
         """
         device = "cuda" if (torch.cuda.is_available() and torch.cuda.device_count() > 0) else "cpu"
 
         _functional_test(
             preds=preds,
-            target=target,
+            targets=targets,
             metric_functional=metric_functional,
             sk_metric=sk_metric,
             metric_args=metric_args,
             atol=self.atol,
             device=device,
             fragment_kwargs=fragment_kwargs,
+            input_order=input_order,
+            key=key,
             **kwargs_update,
         )
 
     def run_class_metric_test(
         self,
         ddp: bool,
-        preds: Tensor,
-        target: Tensor,
+        preds: TEXT_METRIC_INPUT,
+        targets: TEXT_METRIC_INPUT,
         metric_class: Metric,
         sk_metric: Callable,
         dist_sync_on_step: bool,
@@ -352,6 +316,8 @@ class MetricTester:
         check_batch: bool = True,
         fragment_kwargs: bool = False,
         check_scriptable: bool = True,
+        input_order: INPUT_ORDER = INPUT_ORDER.PREDS_FIRST,
+        key: str = None,
         **kwargs_update,
     ):
         """Main method that should be used for testing class. Call this inside testing methods.
@@ -359,7 +325,7 @@ class MetricTester:
         Args:
             ddp: bool, if running in ddp mode or not
             preds: torch tensor with predictions
-            target: torch tensor with targets
+            targets: torch tensor with targets
             metric_class: lightning metric class that should be tested
             sk_metric: callable function that is used for comparison
             dist_sync_on_step: bool, if true will synchronize metric state across
@@ -369,9 +335,12 @@ class MetricTester:
                 calculated per batch per device (and not just at the end)
             check_batch: bool, if true will check if the metric is also correctly
                 calculated across devices for each batch (and not just at the end)
-            fragment_kwargs: whether tensors in kwargs should be divided as `preds` and `target` among processes
+            fragment_kwargs: whether tensors in kwargs should be divided as `preds` and `targets` among processes
+            input_order: Define the ordering for the preds and targets positional arguments.
+            key: The key passed onto the `_assert_allclose` to compare the respective metric from the Dict output
+                against the sk_metric.
             kwargs_update: Additional keyword arguments that will be passed with preds and
-                target when running update on the metric.
+                targets when running update on the metric.
         """
         if not metric_args:
             metric_args = {}
@@ -383,7 +352,7 @@ class MetricTester:
                 partial(
                     _class_test,
                     preds=preds,
-                    target=target,
+                    targets=targets,
                     metric_class=metric_class,
                     sk_metric=sk_metric,
                     dist_sync_on_step=dist_sync_on_step,
@@ -393,6 +362,8 @@ class MetricTester:
                     atol=self.atol,
                     fragment_kwargs=fragment_kwargs,
                     check_scriptable=check_scriptable,
+                    input_order=input_order,
+                    key=key,
                     **kwargs_update,
                 ),
                 [(rank, self.poolSize) for rank in range(self.poolSize)],
@@ -404,7 +375,7 @@ class MetricTester:
                 rank=0,
                 worldsize=1,
                 preds=preds,
-                target=target,
+                targets=targets,
                 metric_class=metric_class,
                 sk_metric=sk_metric,
                 dist_sync_on_step=dist_sync_on_step,
@@ -415,13 +386,15 @@ class MetricTester:
                 device=device,
                 fragment_kwargs=fragment_kwargs,
                 check_scriptable=check_scriptable,
+                input_order=input_order,
+                key=key,
                 **kwargs_update,
             )
 
     @staticmethod
     def run_precision_test_cpu(
-        preds: Tensor,
-        target: Tensor,
+        preds: TEXT_METRIC_INPUT,
+        targets: TEXT_METRIC_INPUT,
         metric_module: Metric,
         metric_functional: Callable,
         metric_args: dict = None,
@@ -430,22 +403,22 @@ class MetricTester:
         """Test if a metric can be used with half precision tensors on cpu
         Args:
             preds: torch tensor with predictions
-            target: torch tensor with targets
+            targets: torch tensor with targets
             metric_module: the metric module to test
             metric_functional: the metric functional to test
             metric_args: dict with additional arguments used for class initialization
             kwargs_update: Additional keyword arguments that will be passed with preds and
-                target when running update on the metric.
+                targets when running update on the metric.
         """
         metric_args = metric_args or {}
         _assert_half_support(
-            metric_module(**metric_args), metric_functional, preds, target, device="cpu", **kwargs_update
+            metric_module(**metric_args), metric_functional, preds, targets, device="cpu", **kwargs_update
         )
 
     @staticmethod
     def run_precision_test_gpu(
-        preds: Tensor,
-        target: Tensor,
+        preds: TEXT_METRIC_INPUT,
+        targets: TEXT_METRIC_INPUT,
         metric_module: Metric,
         metric_functional: Callable,
         metric_args: dict = None,
@@ -454,98 +427,50 @@ class MetricTester:
         """Test if a metric can be used with half precision tensors on gpu
         Args:
             preds: torch tensor with predictions
-            target: torch tensor with targets
+            targets: torch tensor with targets
             metric_module: the metric module to test
             metric_functional: the metric functional to test
             metric_args: dict with additional arguments used for class initialization
             kwargs_update: Additional keyword arguments that will be passed with preds and
-                target when running update on the metric.
+                targets when running update on the metric.
         """
         metric_args = metric_args or {}
         _assert_half_support(
-            metric_module(**metric_args), metric_functional, preds, target, device="cuda", **kwargs_update
+            metric_module(**metric_args), metric_functional, preds, targets, device="cuda", **kwargs_update
         )
 
     @staticmethod
     def run_differentiability_test(
-        preds: Tensor,
-        target: Tensor,
+        preds: TEXT_METRIC_INPUT,
+        targets: TEXT_METRIC_INPUT,
         metric_module: Metric,
         metric_functional: Callable,
         metric_args: dict = None,
+        input_order: INPUT_ORDER = INPUT_ORDER.PREDS_FIRST,
+        key: str = None,
     ):
         """Test if a metric is differentiable or not.
 
         Args:
             preds: torch tensor with predictions
-            target: torch tensor with targets
+            targets: torch tensor with targets
             metric_module: the metric module to test
             metric_args: dict with additional arguments used for class initialization
+            input_order: Define the ordering for the preds and targets positional arguments.
+            key: The key passed onto the `_assert_allclose` to compare the respective metric from the Dict output
+                against the sk_metric.
         """
         metric_args = metric_args or {}
         # only floating point tensors can require grad
         metric = metric_module(**metric_args)
-        if preds.is_floating_point():
-            preds.requires_grad = True
-            out = metric(preds[0], target[0])
+        if input_order == INPUT_ORDER.PREDS_FIRST:
+            out = metric(preds[0], targets[0])
+        elif input_order == INPUT_ORDER.TARGETS_FIRST:
+            out = metric(targets[0], preds[0])
 
-            # Check if requires_grad matches is_differentiable attribute
-            _assert_requires_grad(metric, out)
+        # Check if requires_grad matches is_differentiable attribute
+        _assert_requires_grad(metric, out, key=key)
 
-            if metric.is_differentiable:
-                # check for numerical correctness
-                assert torch.autograd.gradcheck(
-                    partial(metric_functional, **metric_args), (preds[0].double(), target[0])
-                )
-
-            # reset as else it will carry over to other tests
-            preds.requires_grad = False
-
-
-class DummyMetric(Metric):
-    name = "Dummy"
-
-    def __init__(self):
-        super().__init__()
-        self.add_state("x", tensor(0.0), dist_reduce_fx=None)
-
-    def update(self):
-        pass
-
-    def compute(self):
-        pass
-
-
-class DummyListMetric(Metric):
-    name = "DummyList"
-
-    def __init__(self):
-        super().__init__()
-        self.add_state("x", [], dist_reduce_fx=None)
-
-    def update(self):
-        pass
-
-    def compute(self):
-        pass
-
-
-class DummyMetricSum(DummyMetric):
-    def update(self, x):
-        self.x += x
-
-    def compute(self):
-        return self.x
-
-
-class DummyMetricDiff(DummyMetric):
-    def update(self, y):
-        self.x -= y
-
-    def compute(self):
-        return self.x
-
-
-class DummyMetricMultiOutput(DummyMetricSum):
-    def compute(self):
-        return [self.x, self.x]
+        if metric.is_differentiable:
+            # check for numerical correctness
+            assert torch.autograd.gradcheck(partial(metric_functional, **metric_args), (preds[0], targets[0]))